Fix bug where layer was ignored when finding peeps for dialog box.
When the same resource was loaded between maps, this resulted in selecting the WRONG peep. For example, the pre-tutorial cutscene when the ships are sink; Hans
appears in both the lazy loaded map for the portal AND the map the player is on.
Sometimes the wrong Hans was selected which screwed with the location dialog.